requirements.txt
alembic==1.7.7
Authlib==1.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cryptography==36.0.2
ecdsa==0.17.0
Flask==2.1.0
Flask-Cors==3.0.10
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
python-dotenv==0.20.0
python-jose==3.3.0
requests==2.27.1
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.35
urllib3==1.26.9
Werkzeug==2.1.1
zipp==3.8.0
